refactor(speech): move ChannelStt out of cpal-gated backends_common (Plan 2 Task 4)
